requirements.txt
uvicorn==0.18.2
starlette==0.20.4
pydantic==1.9.1
pyyaml==6.0
python-magic==0.4.25
